+++ src/Indexing/BibleIndexer.m
@@ -113,7 +113,7 @@ SearchBookSet *searchBookSet;
  @return array of NSDictionaries with search results. 
  the array is autoreleased, the caller has to make sure to retain it if needed.
  */
-- (NSArray *)performSearchOperation:(NSString *)query constrains:(id)constrains maxResults:(int)maxResults {
+- (NSArray *)performSearchOperation:(NSString *)query constrains:(id)constrains maxResults:(long)maxResults {
     NSMutableArray *array = nil;
     
     [accessLock lock];
